LU-16152 utils: fix integer overflow in cYAML parser
Convert double to int64 correctly in cYAML parser.

Test-Parameters: trivial testlist=sanity env=ONLY=65p
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/48649
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>

+++ lnet/utils/lnetconfig/cyaml.c
@@ -391,7 +391,7 @@ static bool parse_number(struct cYAML *item, const char *input)
 	n = sign * n * pow(10.0, (scale + subscale * signsubscale));
 
 	item->cy_valuedouble = n;
-	item->cy_valueint = (int)n;
+	item->cy_valueint = (int64_t)n;
 	item->cy_type = CYAML_TYPE_NUMBER;
 
 	return true;
